Added integration of Pico SDK lock_core overrides
Slight modification to the lock_core.h header from the pico SDK
was required to override the core macros.

Currently, the default pico sleep implementation is used; however,
this doesn't allow the process ID to be passed from the wait to
the notify handlers. This can be resolved disabling the default
alarm pool and implementing an alarm pool for the scheduler. This
was not tried in this commit.