Disabled default alarm pool on pico SDK integration
This allows use of the sync_internal_yield_until_before macro
in pico SDK Pico Time library instead of default wait/notify.
Swapkernel is then able to handle the wait/notify itself using
Pico's alarms.

The Pico SDK stdio_usb.c is dependent on the default alarm pool.
This can be resolved by removing the offending function with
pre-processor directives and adding an alarm pool just for that
library if the default pool is disabled.