Added multicore support to swapkernel
Multicore with SMP can be enabled by building with
SWAPK_HARDWARE_THREADS greater than 1 (the actual number should be
equal to the number of cores or processor threads). Note that setting
this >1 will likely cause crashes early in the application.

API was changed considerably, and the pico example integration was
updated to reflect that. A simple system call framework was created
for kernel interfacing. Currently this is only used to signal cores
that the scheduler is available.

For the pico SDK example, USB stdout interupts seem to interfere
with processes waking from sleep and will cause a crash after their
context frame is swapped in.