Added unmanaged proc interface
Swapkernel would frequently starve on the Pico platform due to
the lockmaps assigning mutexes created by the USB stdio interface
to the current process and then context-switching to that process.

Now the core provides storage for unmanaged processes, which can
be used to assign USB primatives to an existing process that will
never be executed by Swapkernel.